requirements.txt
aiogram==3.7.0
motor==3.4.0
pymongo==4.6.3
python-dotenv==1.0.1
